requirements.txt
streamlit
pandas
neuralprophet
sentence-transformers
chromadb
httpx
python-dotenv
transformers
torch
llama-cpp-python
